Added property for continuous marquee
Added two new properties : marqueeType and continuousMarqueeSeparator.
The first one is chosen in an enum(LeftRight, Continuous). When set to
Continuous, the label will loop indefinitely separated by itself by the
continuousMarqueeSeparator NSString. LeftRight is the default behavior.
Can be refactored to avoid block duplication in the setText method.
Updated the Demo accordingly.